Refactor enums and combat flags Rename zombie_damage_flags for parity with original naming conventions (e.g., ignore_accessory_2 to bypasses_shield, disable_ballon_pop to doesnt_leave_body). Update scene_type, zombie_type, and projectile_type to use int32_t and include missing variants (Mushroom Garden, Greenhouse, etc.). Refactor attack_flags to use bit-shifts for better readability. Rename projectile motion cattail to homing and add support for bee motion types. Standardize mathematical expressions and JSON serialization across the system to reflect these changes.

+++ object/scene.h
@@ -27,7 +27,7 @@ struct grid_plant_status {
         coffee_bean(nullptr) {}
 };
 
-enum class scene_type : std::uint8_t {
+enum class scene_type : std::int32_t {
     day = 0x0,        // 白天
     night = 0x1,      // 夜晚
     pool = 0x2,       // 泳池
